% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.QP1.R
\name{solve.QP1}
\alias{solve.QP1}
\title{creates QP and negative}
\usage{
\method{solve}{QP1}(Dmat, dvec, Amat, bvec, meq = 0, factorized = FALSE)
}
\arguments{
\item{r}{numeric variable}
}
\description{
creates vimal and negative
}
